fix: migrate formula/symbol/chart extractors to modern ProjectConfig
