+++ CorpusBuilderApp/shared_tools/processors/chart_image_extractor.py
@@ -682,7 +682,7 @@ def main():
     
     if args.project_config:
         # Use project config
-        from shared_tools.config.project_config import ProjectConfig
+        from shared_tools.project_config import ProjectConfig
         project = ProjectConfig.load(args.project_config)
         results = run_with_project_config(project, args.verbose)
     else:
