+++ CorpusBuilderApp/shared_tools/processors/finacial_symbol_processor.py
@@ -791,7 +791,7 @@ def main():
     
     if args.project_config:
         # Use project config
-        from shared_tools.config.project_config import ProjectConfig
+        from shared_tools.project_config import ProjectConfig
         project = ProjectConfig.load(args.project_config)
         results = run_with_project_config(project, args.verbose)
     else:
